Rename Github Pages tasks
